requirements.txt.txt
Flask
flask-cors
gunicorn
google-generativeai
python-dotenv
moviepy
pexels-api
gTTS
requests